pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-lint-cli"
version = "0.2.0"
description = "Quality and security platform for AI agents. Validate MCP servers, scan for vulnerabilities, ship reliable agents."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "agent-lint contributors" },
]
keywords = ["mcp", "ai", "agents", "security", "validation", "lint", "testing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Topic :: Security",
    "Framework :: Pytest",
]
dependencies = [
    "click>=8.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.22",
]
testing = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
agent-lint = "agent_lint.cli:cli"

[project.entry-points."pytest11"]
agent_lint = "agent_lint.ci.pytest_plugin"

[project.urls]
Homepage = "https://github.com/JDoornink/agent-lint"
Issues = "https://github.com/JDoornink/agent-lint/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_lint"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
